+++ exonum-java-binding/core/rust/src/storage/proof_entry.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use exonum_merkledb::{access::FromAccess, Entry, Fork, ObjectHash, Snapshot};
+use exonum_merkledb::{access::FromAccess, Fork, ObjectHash, ProofEntry, Snapshot};
 use jni::{
     objects::{JClass, JObject, JString},
     sys::{jboolean, jbyteArray},
@@ -25,7 +25,7 @@ use handle::{self, Handle};
 use storage::db::{Value, View, ViewRef};
 use utils;
 
-type Index<T> = Entry<T, Value>;
+type Index<T> = ProofEntry<T, Value>;
 
 enum IndexType {
     SnapshotIndex(Index<&'static dyn Snapshot>),
@@ -34,7 +34,7 @@ enum IndexType {
 
 /// Returns pointer to the created `Entry` object.
 #[no_mangle]
-pub extern "system" fn Java_com_exonum_binding_core_storage_indices_EntryIndexProxy_nativeCreate(
+pub extern "system" fn Java_com_exonum_binding_core_storage_indices_ProofEntryIndexProxy_nativeCreate(
     env: JNIEnv,
     _: JClass,
     name: JString,
@@ -58,7 +58,7 @@ pub extern "system" fn Java_com_exonum_binding_core_storage_indices_EntryIndexPr
 
 /// Destroys the underlying `Entry` object and frees memory.
 #[no_mangle]
-pub extern "system" fn Java_com_exonum_binding_core_storage_indices_EntryIndexProxy_nativeFree(
+pub extern "system" fn Java_com_exonum_binding_core_storage_indices_ProofEntryIndexProxy_nativeFree(
     env: JNIEnv,
     _: JClass,
     entry_handle: Handle,
@@ -68,7 +68,7 @@ pub extern "system" fn Java_com_exonum_binding_core_storage_indices_EntryIndexPr
 
 /// Returns the value or null pointer if it is absent.
 #[no_mangle]
-pub extern "system" fn Java_com_exonum_binding_core_storage_indices_EntryIndexProxy_nativeGet(
+pub extern "system" fn Java_com_exonum_binding_core_storage_indices_ProofEntryIndexProxy_nativeGet(
     env: JNIEnv,
     _: JObject,
     entry_handle: Handle,
@@ -88,7 +88,7 @@ pub extern "system" fn Java_com_exonum_binding_core_storage_indices_EntryIndexPr
 
 /// Returns `true` if the entry contains the value.
 #[no_mangle]
-pub extern "C" fn Java_com_exonum_binding_core_storage_indices_EntryIndexProxy_nativeIsPresent(
+pub extern "C" fn Java_com_exonum_binding_core_storage_indices_ProofEntryIndexProxy_nativeIsPresent(
     env: JNIEnv,
     _: JObject,
     entry_handle: Handle,
@@ -104,7 +104,7 @@ pub extern "C" fn Java_com_exonum_binding_core_storage_indices_EntryIndexProxy_n
 
 /// Returns the hash of the value or default hash if value is absent.
 #[no_mangle]
-pub extern "system" fn Java_com_exonum_binding_core_storage_indices_EntryIndexProxy_nativeGetHash(
+pub extern "system" fn Java_com_exonum_binding_core_storage_indices_ProofEntryIndexProxy_nativeGetHash(
     env: JNIEnv,
     _: JObject,
     entry_handle: Handle,
@@ -123,7 +123,7 @@ pub extern "system" fn Java_com_exonum_binding_core_storage_indices_EntryIndexPr
 
 /// Inserts value to the entry.
 #[no_mangle]
-pub extern "system" fn Java_com_exonum_binding_core_storage_indices_EntryIndexProxy_nativeSet(
+pub extern "system" fn Java_com_exonum_binding_core_storage_indices_ProofEntryIndexProxy_nativeSet(
     env: JNIEnv,
     _: JObject,
     entry_handle: Handle,
@@ -144,7 +144,7 @@ pub extern "system" fn Java_com_exonum_binding_core_storage_indices_EntryIndexPr
 
 /// Removes a value from the entry.
 #[no_mangle]
-pub extern "C" fn Java_com_exonum_binding_core_storage_indices_EntryIndexProxy_nativeRemove(
+pub extern "C" fn Java_com_exonum_binding_core_storage_indices_ProofEntryIndexProxy_nativeRemove(
     env: JNIEnv,
     _: JObject,
     entry_handle: Handle,