+++ exonum-java-binding/cryptocurrency-demo/src/test/java/com/exonum/binding/cryptocurrency/CryptocurrencySchemaIntegrationTest.java
@@ -24,12 +24,10 @@ import static com.exonum.binding.cryptocurrency.PredefinedServiceParameters.arti
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.exonum.binding.common.crypto.PublicKey;
-import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.core.storage.database.Snapshot;
 import com.exonum.binding.test.RequiresNativeLibrary;
 import com.exonum.binding.testkit.TestKit;
 import com.exonum.binding.testkit.TestKitExtension;
-import com.google.common.collect.ImmutableList;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
@@ -45,17 +43,6 @@ class CryptocurrencySchemaIntegrationTest {
 
   private static final PublicKey WALLET_OWNER_KEY =
       PredefinedOwnerKeys.FIRST_OWNER_KEY_PAIR.getPublicKey();
-
-  @Test
-  void getStateHashes(TestKit testKit) {
-    Snapshot view = testKit.getSnapshot();
-    CryptocurrencySchema schema = new CryptocurrencySchema(view, SERVICE_NAME);
-
-    HashCode walletsMerkleRoot = schema.wallets().getIndexHash();
-    ImmutableList<HashCode> expectedHashes = ImmutableList.of(walletsMerkleRoot);
-
-    assertThat(schema.getStateHashes()).isEqualTo(expectedHashes);
-  }
 
   @Test
   void walletHistoryNoRecords(TestKit testKit) {