+++ exonum-java-binding/cryptocurrency-demo/src/test/java/com/exonum/binding/cryptocurrency/TransferTxIntegrationTest.java
@@ -43,6 +43,7 @@ import com.exonum.binding.testkit.TestKit;
 import com.exonum.binding.testkit.TestKitExtension;
 import com.exonum.core.messages.Runtime.ExecutionStatus;
 import java.util.Optional;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -119,6 +120,7 @@ class TransferTxIntegrationTest {
   }
 
   @Test
+  @Disabled("ECR-4014")
   @RequiresNativeLibrary
   void executeTransfer_NoSuchFromWallet(TestKit testKit) {
     // Create a receiver’s wallet with the given initial balance
@@ -141,6 +143,7 @@ class TransferTxIntegrationTest {
   }
 
   @Test
+  @Disabled("ECR-4014")
   @RequiresNativeLibrary
   void executeTransfer_NoSuchToWallet(TestKit testKit) {
     // Create a receiver’s wallet with the given initial balance
@@ -163,6 +166,7 @@ class TransferTxIntegrationTest {
   }
 
   @Test
+  @Disabled("ECR-4014")
   @RequiresNativeLibrary
   void executeTransfer_RejectsSameSenderAndReceiver(TestKit testKit) {
     long seed = 1L;
@@ -179,6 +183,7 @@ class TransferTxIntegrationTest {
   }
 
   @Test
+  @Disabled("ECR-4014")
   @RequiresNativeLibrary
   void executeTransfer_InsufficientFunds(TestKit testKit) {
     // Create source and target wallets with the same initial balance