+++ exonum-java-binding/qa-service/src/main/java/com/exonum/binding/qaservice/QaServiceImpl.java
@@ -33,10 +33,9 @@ import com.exonum.binding.core.service.BlockCommittedEvent;
 import com.exonum.binding.core.service.Configuration;
 import com.exonum.binding.core.service.Node;
 import com.exonum.binding.core.storage.database.Fork;
-import com.exonum.binding.core.storage.database.Snapshot;
 import com.exonum.binding.core.storage.database.View;
-import com.exonum.binding.core.storage.indices.EntryIndexProxy;
 import com.exonum.binding.core.storage.indices.MapIndex;
+import com.exonum.binding.core.storage.indices.ProofEntryIndexProxy;
 import com.exonum.binding.core.storage.indices.ProofMapIndexProxy;
 import com.exonum.binding.core.transaction.RawTransaction;
 import com.exonum.binding.core.transaction.Transaction;
@@ -54,7 +53,6 @@ import com.google.protobuf.ByteString;
 import io.vertx.ext.web.Router;
 import java.nio.charset.StandardCharsets;
 import java.time.ZonedDateTime;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import javax.annotation.Nullable;
@@ -101,15 +99,6 @@ public final class QaServiceImpl extends AbstractService implements QaService {
   @Override
   protected QaSchema createDataSchema(View view) {
     return new QaSchema(view, getName());
-  }
-
-  @Override
-  public List<HashCode> getStateHashes(Snapshot snapshot) {
-    List<HashCode> stateHashes = super.getStateHashes(snapshot);
-    // Log the state hashes, so that the values passed to the native part of the framework
-    // are known.
-    logger.info("state hashes: {}", stateHashes);
-    return stateHashes;
   }
 
   @Override
@@ -247,7 +236,7 @@ public final class QaServiceImpl extends AbstractService implements QaService {
   public Optional<ZonedDateTime> getTime() {
     return node.withSnapshot(s -> {
       TimeSchema timeOracle = createDataSchema(s).timeSchema();
-      EntryIndexProxy<ZonedDateTime> currentTime = timeOracle.getTime();
+      ProofEntryIndexProxy<ZonedDateTime> currentTime = timeOracle.getTime();
       return currentTime.toOptional();
     });
   }