+++ exonum-java-binding/CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   See `ProofMapIndexProxy#getProof` and `MapProof`;
   `ProofListIndexProxy.getProof`, `ProofListIndexProxy.getRangeProof` and
   `ListProof`.
+- `ProofEntryIndexProxy` collection.
 - Transaction methods now accept protobuf messages as transaction arguments
   type. (#1304)
 
@@ -32,6 +33,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Removed
 - Classes supporting no longer used tree-like list proof representation.
+- `Schema#getStateHashes` and `Service#getStateHashes` methods. Framework
+  automatically aggregates state hashes of the Merkelized collections.
 
 ## 0.9.0-rc2 - 2019-12-17